+++ contracts/hackatom/src/contract.rs
@@ -314,7 +314,7 @@ mod tests {
 
     #[test]
     fn proper_initialization() {
-        let mut deps = mock_dependencies(&[]);
+        let mut deps = mock_dependencies();
 
         let verifier = String::from("verifies");
         let beneficiary = String::from("benefits");
@@ -342,7 +342,7 @@ mod tests {
 
     #[test]
     fn instantiate_and_query() {
-        let mut deps = mock_dependencies(&[]);
+        let mut deps = mock_dependencies();
 
         let verifier = String::from("verifies");
         let beneficiary = String::from("benefits");
@@ -362,7 +362,7 @@ mod tests {
 
     #[test]
     fn migrate_verifier() {
-        let mut deps = mock_dependencies(&[]);
+        let mut deps = mock_dependencies();
 
         let verifier = String::from("verifies");
         let beneficiary = String::from("benefits");
@@ -394,7 +394,7 @@ mod tests {
 
     #[test]
     fn sudo_can_steal_tokens() {
-        let mut deps = mock_dependencies(&[]);
+        let mut deps = mock_dependencies();
 
         let verifier = String::from("verifies");
         let beneficiary = String::from("benefits");
@@ -437,7 +437,7 @@ mod tests {
 
     #[test]
     fn execute_release_works() {
-        let mut deps = mock_dependencies(&[]);
+        let mut deps = mock_dependencies();
 
         // initialize the store
         let creator = String::from("creator");
@@ -483,7 +483,7 @@ mod tests {
 
     #[test]
     fn execute_release_fails_for_wrong_sender() {
-        let mut deps = mock_dependencies(&[]);
+        let mut deps = mock_dependencies();
 
         // initialize the store
         let creator = String::from("creator");
@@ -528,7 +528,7 @@ mod tests {
     #[test]
     #[should_panic(expected = "This page intentionally faulted")]
     fn execute_panic() {
-        let mut deps = mock_dependencies(&[]);
+        let mut deps = mock_dependencies();
 
         // initialize the store
         let verifier = String::from("verifies");
@@ -555,7 +555,7 @@ mod tests {
 
     #[test]
     fn execute_user_errors_in_api_calls() {
-        let mut deps = mock_dependencies(&[]);
+        let mut deps = mock_dependencies();
 
         let instantiate_msg = InstantiateMsg {
             verifier: String::from("verifies"),
@@ -580,7 +580,7 @@ mod tests {
         // the test framework doesn't handle contracts querying contracts yet,
         // let's just make sure the last step looks right
 
-        let deps = mock_dependencies(&[]);
+        let deps = mock_dependencies();
         let contract = Addr::unchecked("my-contract");
         let bin_contract: &[u8] = b"my-contract";