+++ contracts/ibc-reflect-send/src/contract.rs
@@ -209,7 +209,7 @@ mod tests {
 
     #[test]
     fn instantiate_works() {
-        let mut deps = mock_dependencies(&[]);
+        let mut deps = mock_dependencies();
         let msg = InstantiateMsg {};
         let info = mock_info(CREATOR, &[]);
         let res = instantiate(deps.as_mut(), mock_env(), info, msg).unwrap();