+++ contracts/queue/src/contract.rs
@@ -221,12 +221,12 @@ fn query_list(deps: Deps) -> ListResponse {
 mod tests {
     use super::*;
     use cosmwasm_std::testing::{
-        mock_dependencies, mock_env, mock_info, MockApi, MockQuerier, MockStorage,
+        mock_dependencies_with_balance, mock_env, mock_info, MockApi, MockQuerier, MockStorage,
     };
     use cosmwasm_std::{coins, from_binary, OwnedDeps};
 
     fn create_contract() -> (OwnedDeps<MockStorage, MockApi, MockQuerier>, MessageInfo) {
-        let mut deps = mock_dependencies(&coins(1000, "earth"));
+        let mut deps = mock_dependencies_with_balance(&coins(1000, "earth"));
         let info = mock_info("creator", &coins(1000, "earth"));
         let res = instantiate(deps.as_mut(), mock_env(), info.clone(), InstantiateMsg {}).unwrap();
         assert_eq!(0, res.messages.len());